bot/pyproject.toml
[tool.poetry]
name = "python-telegram-bot"
version = "0.1.0"
description = "A simple Telegram bot built with python-telegram-bot library"
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^13.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"